% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossTalkSig.R
\name{CrossTalkSig}
\alias{CrossTalkSig}
\title{Find significant crosstalk between pairwise clusters}
\usage{
CrossTalkSig(clu_pairs = NULL, per_num = 1000, pvalue = 0.05)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{per_num}{Number of repeat times for permutation test. Default is 1000}

\item{pvalue}{Include the significantly cell-cell communications
with this cutoff of p value from permutation test. Default is \code{0.05}.}
}
\value{
The same list containing the significant cell-cell communications stored
in \code{clu_clu_num} and in \code{clu_clu_score}. For p value, 0 means p < 1/per_num, namely p < 0.001 by default.
}
\description{
Find significant crosstalk between pairwise clusters with permutation test, which might take a long time.
}
\examples{
clu_pairs <- CrossTalkSig(clu_pairs = FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse'))
}
