% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_lrpairs}
\alias{find_lrpairs}
\title{Find highly expressed ligand-receptor pairs}
\usage{
find_lrpairs(
  object,
  lrpairs,
  min_cell_num = 10,
  cell_min_pct = 0.1,
  p_value = 0.05
)
}
\arguments{
\item{object}{scCrossTalk object after \code{\link{create_scCrossTalk}}}

\item{lrpairs}{A data.frame of the system data containing ligand-receptor pairs of \code{'Human'} and \code{'Mouse'}. see \code{\link{demo_lrpairs}}}

\item{min_cell_num}{Min cell number for each cell type. Default is \code{10}}

\item{cell_min_pct}{Include the ligand and receptor gene detected in at least this many cells in each cluster. Default is \code{0.1}}

\item{p_value}{Include the significantly highly expressed ligands and receptors with this cutoff of p value from Z score. Default is \code{0.05}}
}
\value{
scCrossTalk object containing the enriched ligand-receptor pairs and cell-cell communications
}
\description{
Find highly expressed ligands and receptors between pairwise cell types using Z score
}
