% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{find_pathways}
\alias{find_pathways}
\title{Get pathways}
\usage{
find_pathways(target_genes, pathways, species)
}
\arguments{
\item{target_genes}{Character of one or more target genes}

\item{pathways}{A data.frame of the system data containing gene-gene interactions and pathways from KEGG and Reactome for \code{'Human'}, \code{'Mouse'} or \code{'Rat'}. see \code{\link{demo_pathways}}}

\item{species}{A character meaning species of the target genes.\code{'Human'}, \code{'Mouse'} or \code{'Rat'}}
}
\value{
Pathways for one or more target genes
}
\description{
Get pathways for target genes
}
