% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_top_lrpairs}
\alias{get_top_lrpairs}
\title{Get top n LR-pairs}
\usage{
get_top_lrpairs(object, top_n = 10)
}
\arguments{
\item{object}{scCrossTalk object after \code{\link{create_scCrossTalk}}}

\item{top_n}{Number of top LR-pairs. Default is {10}}
}
\value{
A data.frame containing the top n LR-pairs for each cell pairs
}
\description{
Get top n LR-pairs for each cell pairs
}
