% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{lrpairs}
\alias{lrpairs}
\title{lrpairs}
\format{
An object of class \code{data.frame} with 5427 rows and 3 columns.
}
\source{
\url{http://tcm.zju.edu.cn/celltalkdb/}
}
\usage{
lrpairs
}
\description{
A data.frame containing ligand-receptor interactions for \code{"Human"} and \code{"Mouse"} from CellTalkDB.
}
\keyword{datasets}
