% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cci_circle}
\alias{plot_cci_circle}
\title{Circle plot of cell-cell communications}
\usage{
plot_cci_circle(
  object,
  celltype = NULL,
  ligand = NULL,
  receptor = NULL,
  celltype_color = NULL,
  edge_color = NULL,
  edge_type = "fan",
  show_type = "number",
  if_show_autocrine = FALSE,
  edge_alpha = 0.5,
  node_size = 10,
  text_size = 5
)
}
\arguments{
\item{object}{scCrossTalk object after \code{\link{find_lrpairs}}}

\item{celltype}{which cell types to plot. Default is to plot all cell types}

\item{ligand}{which ligand to use. Default is to plot all inferred ligands}

\item{receptor}{which receptor to use. Default is to plot all inferred receptors}

\item{celltype_color}{Colors for the cell types, whose length must be equal to \code{celltype}}

\item{edge_color}{Colors for the edges from the sender cell type, whose length must be equal to \code{celltype}}

\item{edge_type}{Types for the edges. \code{"fan"} by default, \code{"link"}, \code{"hive"}}

\item{show_type}{which type to show, \code{"number"} and \code{"score"} for sum of inferred LR number and score, respectively. Default is \code{"number"}}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{edge_alpha}{Transparency of edge. Default is \code{0.5}}

\item{node_size}{Size of node. Default is \code{10}}

\item{text_size}{Size of text. Default is \code{5}}
}
\value{
Circle plot of cell-cell communications mediated by ligand-receptor interactions
}
\description{
Circle plot of cell-cell communications from senders to receivers with the sum of inferred number or score of ligand-receptor interactions
}
