% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cci_lrpairs_heatmap}
\alias{plot_cci_lrpairs_heatmap}
\title{Heatmap plot of all lrpairs among cell pairs}
\usage{
plot_cci_lrpairs_heatmap(
  object,
  celltype = NULL,
  ligand = NULL,
  receptor = NULL,
  if_show_autocrine = FALSE,
  if_horizontal = TRUE,
  text_size = 10,
  viridis_option = "D",
  ...
)
}
\arguments{
\item{object}{scCrossTalk object after \code{\link{find_lrpairs}}}

\item{celltype}{which cell types to plot. Default is to plot all cell types}

\item{ligand}{which ligand to use. Default is to plot all inferred ligands}

\item{receptor}{which receptor to use. Default is to plot all inferred receptors}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{if_horizontal}{Whether to plot with the horizontal direction. Default is \code{TRUE}}

\item{text_size}{Size of text labels. Default is \code{10}}

\item{viridis_option}{option in \code{\link[viridis]{scale_color_viridis}}, can be "A", "B", "C", "D", "E", "F", "G", "H". Default is "D".}

\item{...}{parameters pass to \code{\link[heatmaply]{heatmaply}}, e.g., grid_color}
}
\value{
Heatmap plot of all lrpairs among cell pairs
}
\description{
heatmap plot of inferred lrpairs. Rows for cell pairs, and columns for LR pairs by default.
}
