% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_cci_sankey}
\alias{plot_cci_sankey}
\title{Sankey plot of cell-cell communications}
\usage{
plot_cci_sankey(
  object,
  celltype = NULL,
  ligand = NULL,
  receptor = NULL,
  celltype_color = NULL,
  edge_color = NULL,
  show_type = "number",
  if_show_autocrine = FALSE,
  edge_alpha = 0.5,
  node_size = 40,
  text_size = 15,
  node_pad = 20,
  ...
)
}
\arguments{
\item{object}{scCrossTalk object after \code{\link{find_lrpairs}}}

\item{celltype}{which cell types to plot. Default is to plot all cell types}

\item{ligand}{which ligand to use. Default is to plot all inferred ligands}

\item{receptor}{which receptor to use. Default is to plot all inferred receptors}

\item{celltype_color}{Colors for the cell types, whose length must be equal to \code{celltype}}

\item{edge_color}{Colors for the edges from the sender cell type, whose length must be equal to \code{celltype}, Or use \code{"NO"} to cancel it}

\item{show_type}{which type to show, \code{"number"} and \code{"score"} for sum of inferred LR number and score, respectively. Default is \code{"number"}}

\item{if_show_autocrine}{Whether to show autocrine. Default is \code{FALSE}}

\item{edge_alpha}{Transparency of edge. Default is \code{0.5}}

\item{node_size}{Size of node. Default is \code{40}}

\item{text_size}{Size of text. Default is \code{15}}

\item{node_pad}{Size of node padding. Numeric essentially influences the width height. Default is \code{20}}

\item{...}{parameters pass to \code{\link[networkD3]{sankeyNetwork}}}
}
\value{
Sankey plot of cell-cell communications mediated by ligand-receptor interactions
}
\description{
Sankey plot of cell-cell communications from senders to receivers with the sum of inferred number or score of ligand-receptor interactions
}
