% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lrpairs_chord}
\alias{plot_lrpairs_chord}
\title{Chord plot of ligand-receptor pairs}
\usage{
plot_lrpairs_chord(
  object,
  celltype_sender,
  celltype_receiver,
  celltype_color = NULL,
  ligand = NULL,
  receptor = NULL,
  edge_color = NULL,
  edge_type = "circle",
  text_size = 0.5,
  y_scale = 1,
  ...
)
}
\arguments{
\item{object}{scCrossTalk object after \code{\link{find_lrpairs}}}

\item{celltype_sender}{Name of celltype_sender. One or more cell types}

\item{celltype_receiver}{Name of celltype_receiver. One or more cell types}

\item{celltype_color}{Colors for the celltype_sender nodes and celltype_receiver nodes, or use \code{"NO"} to make it simple}

\item{ligand}{which ligand to use. Default is to plot all inferred ligands}

\item{receptor}{which receptor to use. Default is to plot all inferred receptors}

\item{edge_color}{Colors for the edges from the sender cell type}

\item{edge_type}{Types for the edges from the sender cell type. Default is \code{"circle"}. \code{"big.arrow"} for big arrow, "triangle" for triangle, "ellipse" for ellipse, "curved" for curved. Details see \code{\link[circlize]{chordDiagram}}}

\item{text_size}{Size of text labels. Default is \code{0.5}}

\item{y_scale}{y_scale to adjust the text. Default is \code{1}}

\item{...}{parameters pass to \code{\link[circlize]{chordDiagram}}, e.g., link.arr.width, link.arr.length, link.arr.col}
}
\value{
Chord plot of ligand-receptor pairs from senders to receivers
}
\description{
Chord plot of ligand-receptor pairs from senders to receivers.
}
