% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_lrpairs_heatmap}
\alias{plot_lrpairs_heatmap}
\title{Heatmap plot of ligand-receptor pairs}
\usage{
plot_lrpairs_heatmap(
  object,
  celltype_sender,
  celltype_receiver,
  ligand = NULL,
  receptor = NULL,
  text_size = 5,
  if_horizontal = TRUE,
  viridis_option = "D",
  ...
)
}
\arguments{
\item{object}{scCrossTalk object after \code{\link{find_lrpairs}}}

\item{celltype_sender}{Name of celltype_sender. One or more cell types}

\item{celltype_receiver}{Name of celltype_receiver. One or more cell types}

\item{ligand}{which ligand to use. Default is to plot all inferred ligands}

\item{receptor}{which receptor to use. Default is to plot all inferred receptors}

\item{text_size}{Size of text labels. Default is \code{3}}

\item{if_horizontal}{Whether to plot with the horizontal direction. Default is \code{TRUE}}

\item{viridis_option}{option in \code{\link[viridis]{scale_color_viridis}}, can be "A", "B", "C", "D", "E", "F", "G", "H". Default is "D".}

\item{...}{parameters pass to \code{\link[heatmaply]{heatmaply}}, e.g., grid_color}
}
\value{
Heatmap plot of ligand-receptor pairs
}
\description{
Heatmap plot of ligand-receptor pairs from senders to receivers.
}
