% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCrossTalkCircle.R
\name{PlotCrossTalkCircle}
\alias{PlotCrossTalkCircle}
\title{Plot crosstalk between pairwise clusters}
\usage{
PlotCrossTalkCircle(clu_pairs = NULL, show_type = NULL, show_sig = F)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{show_type}{Which to show, sum of LR pairs number or LR score.
Input 'number' to show sum of LR pairs number; Input 'score' to show sum of LR pairs score}

\item{show_sig}{Whether to show significant crosstalk between pairwise clusters. Default is FALSE
To show significant crosstalk, please run \code{\link{CrossTalkSig}}}
}
\value{
Circle plot of LR pairs between pairwise clusters
}
\description{
Plot crosstalk between pairwise clusters by sum of LR pairs number or score between pairwise clusters with Circle plot
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat, species = 'Mouse')

PlotCrossTalkCircle(clu_pairs = clu_pairs, show_type = 'number')

PlotCrossTalkCircle(clu_pairs = clu_pairs, show_type = 'score', show_sig = T)
}
