% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCrossTalkHeat.R
\name{PlotCrossTalkHeat}
\alias{PlotCrossTalkHeat}
\title{Plot crosstalk between pairwise clusters}
\usage{
PlotCrossTalkHeat(clu_pairs = NULL, show_type = NULL,
  if_directed = T, show_sig = F, color_low = "white",
  color_high = "red", border_color = "grey60", cluster_rows = T,
  cluster_cols = T, symbol = "*", symbol_col = "black",
  symbol_size = 12)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{show_type}{Which to show, sum of LR pairs number or LR score
Input 'number' to show sum of LR pairs number; Input 'score' to show sum of LR pairs score}

\item{if_directed}{Considering the direction of LR pairs between two clusters. Default is TRUE}

\item{show_sig}{Whether to show significant crosstalk between pairwise clusters. Default is FALSE
To show significant crosstalk, please run \code{\link{FindCrossTalk}}}

\item{color_low}{Color of the lowest score. Default is 'white'}

\item{color_high}{Color of the highest score. Default is 'red'}

\item{border_color}{Color of cell borders on heatmap, use NA if no border should be drawn.Default is 'grey60'}

\item{cluster_rows}{Boolean values determining if rows should be clustered. Default is TRUE}

\item{cluster_cols}{Boolean values determining if columns should be clustered. Default is TRUE}

\item{symbol}{Significant symbol to plot. Default is '*'}

\item{symbol_col}{Color of the significant symbol. Default is 'black'}

\item{symbol_size}{Size of the significant symbol. Default is 12}
}
\value{
A corrplot containing sum of LR pairs number or scores. When
if_directed is TRUE, rows mean the source(ligand),columns means the target(receptor)
}
\description{
Plot crosstalk between pairwise clusters by sum of LR pairs number and scores for pairwise clusters with heatmap
}
\examples{
# Directed plot
plotpairsCor(clu_pairs = FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse'),
             if_directed = T)

# Undirected plot
plotpairsCor(clu_pairs = FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse'),
             if_directed = F)
}
