% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCrossTalkNet.R
\name{PlotCrossTalkNet}
\alias{PlotCrossTalkNet}
\title{Plot crosstalk between pairwise clusters}
\usage{
PlotCrossTalkNet(clu_pairs = NULL, show_type = NULL,
  layout = "nicely", show_sig = F, edge_col = "black",
  edge_alpha = 0.1, node_size_min = 5, node_size_max = 10,
  text_size = 3, text_col = "black")
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{show_type}{Which to show, sum of LR pairs number or LR score.
Input 'number' to show sum of LR pairs number; Input 'score' to show sum of LR pairs score}

\item{layout}{Layout of net plot, e.g., 'fr','mds',''randomly','dh',
'gem','graphopt','grid','sphere','kk','lgl'. Default is 'nicely'}

\item{show_sig}{Whether to show significant crosstalk between pairwise clusters. Default is FALSE
To show significant crosstalk, please run \code{\link{CrossTalkSig}}}

\item{edge_col}{Color of edge. Default is 'black'}

\item{edge_alpha}{Transparency of edge. Default is 0.1}

\item{node_size_min}{Minimum size of node. Default is 5}

\item{node_size_max}{Maximum size of node. Default is 10}

\item{text_size}{Size of text labels. Default is 3}

\item{text_col}{Color of text labels Default is 'black'}
}
\value{
Sankey plot of LR pairs between pairwise clusters
}
\description{
Plot crosstalk between pairwise clusters by sum of LR pairs number or score between pairwise clusters with net plot
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat, species = 'Mouse')

PlotCrossTalkNet(clu_pairs = clu_pairs, show_type = 'number')

PlotCrossTalkNet(clu_pairs = clu_pairs, show_type = 'score', show_sig = T)
}
