% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPairsBubble.R
\name{PlotPairsBubble}
\alias{PlotPairsBubble}
\title{Plot LR pairs between pairwise clusters}
\usage{
PlotPairsBubble(
  clu_pairs = NULL,
  LR_pairs = "all",
  show_clusters = "all",
  if_directed = T,
  show_sig = F,
  bubble_col = "black",
  bubble_alpha = 0.6,
  bubble_max_size = 20,
  show_text_cutoff = 1
)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{LR_pairs}{Which LR pairs to show. Default is 'all'. Or custom LR pairs, e.g., App_Tspan12, App_Notch2}

\item{show_clusters}{Which clusters related LR pairs to show. Default is 'all'}

\item{if_directed}{Considering the direction of LR pairs between two clusters. Default is TRUE}

\item{show_sig}{Whether to show significant highly expressed LR pairs between pairwise clusters. Default is FALSE
To show significant LR pairs, please run \code{\link{PairsSig}}}

\item{bubble_col}{Color of bubble. Default is 'black'}

\item{bubble_alpha}{Transparency of bubble. Default is 0.6}

\item{bubble_max_size}{Maximum size of bubble. Default is 20}

\item{show_text_cutoff}{Cutoff to show text label (LR score). Default is 0, namely to show all texts}
}
\value{
A bubble plot of LR pairs for pairwise clusters
}
\description{
Plot LR pairs between pairwise clusters with a bubble plot of LR pairs for pairwise clusters
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse')

# Plot all LR pairs related with cluster 1
PlotPairsBubble (clu_pairs = clu_pairs,
                 LR_pairs = 'all',
                 show_clusters = '1')
}
