% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPairsCircle.R
\name{PlotPairsCircle}
\alias{PlotPairsCircle}
\title{Plot LR pairs between pairwise clusters}
\usage{
PlotPairsCircle(clu_pairs = NULL, show_sig = F, ligand_clu = NULL,
  receptor_clu = NULL)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{show_sig}{Whether to show significant highly expressed LR pairs between pairwise clusters. Default is FALSE
To show significant LR pairs, please run \code{\link{PairsSig}}}

\item{ligand_clu}{Cluster of the ligand, e.g., '1'}

\item{receptor_clu}{Cluster of the receptor, e.g., '2'}
}
\value{
A circle plot of LR pairs between pairwise clusters
}
\description{
Plot LR pairs between pairwise clusters with a circle plot of LR pairs between pairwise clusters
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse')

PlotPairsViolin(clu_pairs = clu_pairs,
                show_sig = T,
                ligand_clu = '1',
                receptor_clu = '2')
}
