% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPairsDim.R
\name{PlotPairsDim}
\alias{PlotPairsDim}
\title{Plot a LR pair between pairwise clusters}
\usage{
PlotPairsDim(clu_pairs = NULL, ligand = NULL, ligand_clu = NULL,
  receptor = NULL, receptor_clu = NULL, reduction = "umap",
  size = 1, text_size = 12)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{ligand}{Ligand gene name, e.g., 'App'.}

\item{ligand_clu}{Cluster of the ligand, e.g., '1'}

\item{receptor}{Receptor gene name, e.g., 'Tspan12'.}

\item{receptor_clu}{Cluster of the receptor, e.g., '2'}

\item{reduction}{Which dimensionality reduction to use, umap, tsne, pca.}

\item{size}{Size of the point. Default is 1}

\item{text_size}{Size of the axis x and y text.}
}
\value{
A point plot of gene expression for pairwise ligands and receptors
}
\description{
Plot a LR pair between pairwise clusters with a point plot of gene expression for pairwise ligands and receptors
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse')

PlotPairsDim(clu_pairs = clu_pairs,
             ligand = 'App',
             ligand_clu = '1',
             receptor = 'Tspan12',
             receptor_clu = '2')
}
