% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPairsNode.R
\name{PlotPairsNode}
\alias{PlotPairsNode}
\title{Plot LR pairs between pairwise clusters}
\usage{
PlotPairsNode(clu_pairs = NULL, show_sig = F, edge_width = 1,
  edge_alpha = 0.5, node_size_min = 1, node_size_max = 10,
  text_size = 3)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{show_sig}{Whether to show significant highly expressed LR pairs between pairwise clusters. Default is FALSE
To show significant LR pairs, please run \code{\link{PairsSig}}}

\item{edge_width}{Width of edge. Default is 1}

\item{edge_alpha}{Transparency of edge. Default is 0.5}

\item{node_size_min}{Minimum size of node. Default is 1}

\item{node_size_max}{Maximum size of node. Default is 10}

\item{text_size}{Size of text labels. Default is 3}
}
\value{
A node plot of LR pairs between pairwise clusters. The number behind gene is the cluster,
L, means ligand;R, means receptor. Size of gene node is the mean expression of this gene in this cluster
}
\description{
Plot LR pairs between pairwise clusters with a node plot of LR pairs between pairwise clusters
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse')

PlotPairsNode(clu_pairs = clu_pairs, show_sig = F)
}
