% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPairsViolin.R
\name{PlotPairsViolin}
\alias{PlotPairsViolin}
\title{Plot a LR pair between pairwise clusters}
\usage{
PlotPairsViolin(clu_pairs = NULL, ligand = NULL, ligand_clu = NULL,
  receptor = NULL, receptor_clu = NULL, show_jitter = T,
  jitter_size = 2)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{ligand}{Ligand gene name, e.g., 'App'}

\item{ligand_clu}{Cluster of the ligand, e.g., '1'}

\item{receptor}{Receptor gene name, e.g., 'Tspan12'.}

\item{receptor_clu}{Cluster of the receptor, e.g., '2'}

\item{show_jitter}{to show jitter. Default is TRUE}

\item{jitter_size}{Size of jitter. Default is 2}
}
\value{
A violin plot of gene expression distribution for pairwise ligand and receptor
}
\description{
Plot a LR pair between pairwise clusters with a violin plot of
gene expression distribution for pairwise ligand and receptor
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse')

PlotPairsViolin(clu_pairs = clu_pairs,
                ligand = 'App',
                ligand_clu = '1',
                receptor = 'Tspan12',
                receptor_clu = '2')
}
