% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathway.R
\name{PlotPathway}
\alias{PlotPathway}
\title{Plot a pathway for a cluster}
\usage{
PlotPathway(
  clu_pairs = NULL,
  pathway_name = NULL,
  cluster = NULL,
  layout = "nicely",
  show_text_cutoff = -1,
  node_size = 10,
  text_size = 3,
  text_col = "black",
  edge_width = 0.5,
  edge_col = "black",
  edge_alpha = 0.2
)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{pathway_name}{Name of pathway, e.g., 'VEGFA-VEGFR2 Pathway'}

\item{cluster}{Pathway in which cluster to plot, e.g., '1'}

\item{layout}{Layout of net plot, e.g., 'nicely','fr','mds',''randomly','dh',
'gem','graphopt','grid','sphere','kk','lgl'. Default is 'nicely'}

\item{show_text_cutoff}{Cutoff to show text label. Default is -1, namely to show all texts}

\item{node_size}{Minimum size of node. Default is 10}

\item{text_size}{Size of text labels. Default is 3}

\item{text_col}{Color of text labels. Default is 'black'}

\item{edge_width}{Width of edge. Default is 0.5}

\item{edge_col}{Color of edge.. Default is 'black'}

\item{edge_alpha}{Transparency of edge. Default is 0.2}
}
\value{
A net plot of a pathway containg the expression of genes for a cluster
}
\description{
Plot a pathway for a cluster showing the expression of each gene with a net plot
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse')

PlotPathway(clu_pairs = clu_pairs,
            pathway_name = 'VEGFA-VEGFR2 Pathway',
            cluster = '1')
}
