% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathwayNet.R
\name{PlotPathwayNet}
\alias{PlotPathwayNet}
\title{Plot pathways between pairwise clusters}
\usage{
PlotPathwayNet(clu_pairs = NULL, show_sig = F, show_clu_node = T,
  layout = "nicely", show_text_cutoff = 0, node_size_min = 5,
  node_size_max = 10, text_size = 3, text_col = "black",
  edge_width = 0.5, edge_col = "black", edge_alpha = 0.2)
}
\arguments{
\item{clu_pairs}{A list generated from \code{\link{FindPairs}}}

\item{show_sig}{Whether to show significant highly expressed LR pairs between pairwise clusters. Default is FALSE
To show significant LR pairs, please run \code{\link{PairsSig}}}

\item{show_clu_node}{Whether to show cluster nodes. Default is TRUE}

\item{layout}{Layout of net plot, e.g., 'fr','mds',''randomly','dh',
'gem','graphopt','grid','sphere','kk','lgl'. Default is 'nicely'}

\item{show_text_cutoff}{Cutoff to show text label. Default is 0, namely to show all texts}

\item{node_size_min}{Minimum size of node. Default is 5}

\item{node_size_max}{Maximum size of node. Default is 10}

\item{text_size}{Size of text labels. Default is 3}

\item{text_col}{Color of text labels. Default is 'black'}

\item{edge_width}{Width of edge. Default is 0.5}

\item{edge_col}{Color of edge.. Default is 'black'}

\item{edge_alpha}{Transparency of edge. Default is 0.2}
}
\value{
A net plot of pathways between pairwise clusters.
L, means ligand;R, means receptor. Size of node is the mean expression of this gene in this cluster
}
\description{
Plot pathways between pairwise clusters with a net plot of pathways between pairwise clusters
}
\examples{
clu_pairs <- FindPairs(object = mouse_kidney_203_Seurat,species = 'Mouse')

PlotPathwayNet(clu_pairs = clu_pairs, show_sig = F)
}
