% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_gene.R
\name{rev_gene}
\alias{rev_gene}
\title{Pre-processing step: revising gene symbols}
\usage{
rev_gene(data = NULL, data_type = NULL, species = NULL, geneinfo = NULL)
}
\arguments{
\item{data}{A data.frame or matrix or dgCMatrix}

\item{data_type}{A character to define the type of \code{data}, select \code{'count'} for the data matrix, \code{'lrpairs'} for the lrpairs data.frame}

\item{species}{Species of the data.\code{'Human'} or \code{'Mouse'}}

\item{geneinfo}{A data.frame of the system data containing gene symbols of \code{'Human'} or \code{'Mouse'} updated on June 19, 2022 for revising gene symbols}
}
\value{
A new data.frame, matrix, or dgCMatrix.
}
\description{
Revise genes according to NCBI Gene symbols updated in June 19, 2022 for count matrix and user-custom lrpairs data.frame
}
