% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{scCrossTalk}
\alias{scCrossTalk}
\alias{scCrossTalk-class}
\title{Definition of 'scCrossTalk' class}
\description{
An S4 class containing the data, meta, and results of inferred cell-cell communications mediated by ligand-receptor interactions.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list containing the data.}

\item{\code{meta}}{A data.frame containing the meta data.}

\item{\code{species}}{A character containing the species.}

\item{\code{lrdb}}{A data.frame containing expressed ligand and receptor genes.}

\item{\code{result}}{A list containing enriched ligand-receptor interactions and cell-cell communications.}
}}

